requirements.txt
# Core dependencies
pandas==2.2.2
matplotlib==3.9.0
python-dateutil==2.9.0.post0
dateparser==1.2.0
tqdm==4.66.4
python-dotenv==1.0.1

# LLM and AI frameworks
langchain==0.2.14
langchain-core==0.2.38
langchain-openai==0.1.22
langchain-community==0.2.12
openai==1.37.0

# Hugging Face Transformers for local models
transformers==4.44.2
torch>=2.0.0
